I think I know why the eye-tracker y-pos kept getting shifted:
moving left/right does not cause one side to mirror the other side, because the negativity of one side is just the "relaxing" of the outer muscle--and the relaxation amount is less than the tensing amount of the other side.

So it sees the left/right as having a net positive result, causing my naive code to think the eye was looking up some.

See here for illustration of eye muscles: http://www.innerbody.com/image/nerv08.html